# fix(perf): shortcut getStringWidth for ASCII-only strings (#4776)