### Before
Total run time for 1000 repetitions for `LspLanguageService.js` is **182276ms**, average run is **181.80ms**.
```
$ cat LspLanguageService.js | NODE_ENV=production node --inspect-brk ./bin/prettier.js --stdin-filepath LspLanguageService.js --loglevel debug --debug-repeat 1000 > /dev/null
Debugger listening on ws://127.0.0.1:9229/d085f49d-a12c-432e-ad7e-acf322152bd3
For help see https://nodejs.org/en/docs/inspector
Debugger attached.
[debug] normalized argv: {"color":true,"editorconfig":true,"stdin-filepath":"LspLanguageService.js","loglevel":"debug","debug-repeat":"1000","plugin-search-dir":[],"plugin":[],"ignore-path":".prettierignore","config-precedence":"cli-override","_":[]}
[debug] resolve config from '/Users/ivanbabak/_sompylasar/_github/prettier/LspLanguageService.js'
[debug] loaded options `null`
[debug] applied config-precedence (cli-override): {"filepath":"LspLanguageService.js"}
[debug] '--debug-repeat' option found, running formatWithCursor 1000 times.
[debug] '--debug-repeat' measurements for formatWithCursor: {
[debug]   "repeat": 1000,
[debug]   "hz": 5.500298017789999,
[debug]   "ms": 181.80833052420616
[debug] }
Waiting for the debugger to disconnect...
```

### After
Total run time for 1000 repetitions for `LspLanguageService.js` is **145246ms**, average run is **144.85ms (20.32% speed-up)**.
```
$ cat LspLanguageService.js | NODE_ENV=production node --inspect-brk ./bin/prettier.js --stdin-filepath LspLanguageService.js --loglevel debug --debug-repeat 1000 > /dev/null
Debugger listening on ws://127.0.0.1:9229/1a338c30-c52c-4ae7-8e11-9aac0a2b736a
For help see https://nodejs.org/en/docs/inspector
Debugger attached.
[debug] normalized argv: {"color":true,"editorconfig":true,"stdin-filepath":"LspLanguageService.js","loglevel":"debug","debug-repeat":"1000","plugin-search-dir":[],"plugin":[],"ignore-path":".prettierignore","config-precedence":"cli-override","_":[]}
[debug] resolve config from '/Users/ivanbabak/_sompylasar/_github/prettier/LspLanguageService.js'
[debug] loaded options `null`
[debug] applied config-precedence (cli-override): {"filepath":"LspLanguageService.js"}
[debug] '--debug-repeat' option found, running formatWithCursor 1000 times.
[debug] '--debug-repeat' measurements for formatWithCursor: {
[debug]   "repeat": 1000,
[debug]   "hz": 6.903349947385214,
[debug]   "ms": 144.85720811223985
[debug] }
Waiting for the debugger to disconnect...
```

Refs https://github.com/prettier/prettier/issues/4776